[metasrv] refactor: extract state machine into dir metasrv/raft/state_machine
