+++ metasrv/src/meta_service/cmd.rs
@@ -23,7 +23,7 @@ use common_metatypes::Table;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::meta_service::Node;
+use crate::raft::state_machine::Node;
 
 /// A Cmd describes what a user want to do to raft state machine
 /// and is the essential part of a raft log.
