+++ metasrv/src/meta_service/mod.rs
@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use applied_state::AppliedState;
 pub use cmd::Cmd;
 pub use errors::RetryableError;
 pub use errors::ShutdownError;
 pub use log_entry::LogEntry;
 pub use meta_service_impl::MetaServiceImpl;
 pub use network::Network;
-pub use placement::Placement;
 pub use raft_state_kv::RaftStateKey;
 pub use raft_state_kv::RaftStateValue;
 pub use raft_txid::RaftTxId;
@@ -28,12 +26,6 @@ pub use raft_types::NodeId;
 pub use raft_types::Term;
 pub use raftmeta::MetaNode;
 pub use raftmeta::MetaRaftStore;
-pub use snapshot::Snapshot;
-pub use state_machine::Node;
-pub use state_machine::Slot;
-pub use state_machine::StateMachine;
-pub use state_machine_meta::StateMachineMetaKey;
-pub use state_machine_meta::StateMachineMetaValue;
 
 pub use crate::protobuf::meta_service_client::MetaServiceClient;
 pub use crate::protobuf::meta_service_server::MetaService;
@@ -41,14 +33,14 @@ pub use crate::protobuf::meta_service_server::MetaServiceServer;
 pub use crate::protobuf::GetReply;
 pub use crate::protobuf::GetReq;
 pub use crate::protobuf::RaftMes;
+pub use crate::raft::state_machine::applied_state::AppliedState;
+pub use crate::raft::state_machine::placement::Placement;
 
-pub mod applied_state;
 pub mod cmd;
 pub mod errors;
 pub mod log_entry;
 pub mod meta_service_impl;
 pub mod network;
-pub mod placement;
 pub mod raft_db;
 pub mod raft_log;
 pub mod raft_state;
@@ -56,16 +48,11 @@ pub mod raft_state_kv;
 pub mod raft_txid;
 pub mod raft_types;
 pub mod raftmeta;
-pub mod snapshot;
-pub mod state_machine;
-pub mod state_machine_meta;
 
 #[cfg(test)]
 mod meta_service_impl_test;
 #[cfg(test)]
 mod meta_store_test;
-#[cfg(test)]
-mod placement_test;
 #[cfg(test)]
 mod raft_log_test;
 #[cfg(test)]
@@ -73,8 +60,6 @@ mod raft_state_test;
 #[cfg(test)]
 mod raft_types_test;
 #[cfg(test)]
-mod raftmeta_test;
+pub mod raftmeta_test;
 #[cfg(test)]
-mod state_machine_test;
-#[cfg(test)]
-mod testing;
+pub mod testing;
