+++ metasrv/src/raft/state_machine/sm.rs
@@ -43,18 +43,18 @@ use serde::Serialize;
 use sled::IVec;
 
 use crate::configs;
-use crate::meta_service::placement::rand_n_from_m;
 use crate::meta_service::raft_db::get_sled_db;
-use crate::meta_service::state_machine_meta::StateMachineMetaKey::LastMembership;
 use crate::meta_service::AppliedState;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
 use crate::meta_service::NodeId;
 use crate::meta_service::Placement;
-use crate::meta_service::StateMachineMetaKey;
-use crate::meta_service::StateMachineMetaKey::Initialized;
-use crate::meta_service::StateMachineMetaKey::LastApplied;
-use crate::meta_service::StateMachineMetaValue;
+use crate::raft::state_machine::placement::rand_n_from_m;
+use crate::raft::state_machine::StateMachineMetaKey;
+use crate::raft::state_machine::StateMachineMetaKey::Initialized;
+use crate::raft::state_machine::StateMachineMetaKey::LastApplied;
+use crate::raft::state_machine::StateMachineMetaKey::LastMembership;
+use crate::raft::state_machine::StateMachineMetaValue;
 use crate::sled_store::sled_key_space;
 use crate::sled_store::sled_key_space::StateMachineMeta;
 use crate::sled_store::AsKeySpace;
