+++ metasrv/src/sled_store/sled_tree_test.rs
@@ -22,9 +22,9 @@ use common_runtime::tokio;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
 use crate::meta_service::LogIndex;
-use crate::meta_service::StateMachineMetaKey::Initialized;
-use crate::meta_service::StateMachineMetaKey::LastApplied;
-use crate::meta_service::StateMachineMetaValue;
+use crate::raft::state_machine::StateMachineMetaKey::Initialized;
+use crate::raft::state_machine::StateMachineMetaKey::LastApplied;
+use crate::raft::state_machine::StateMachineMetaValue;
 use crate::sled_store::sled_key_space;
 use crate::sled_store::sled_key_space::SledKeySpace;
 use crate::sled_store::sled_key_space::StateMachineMeta;
